requirement.txt
GitPython==3.1.32
PyYAML==6.0.1
requests==2.31.0
rich==13.4.2
textual==0.29.0

